pyproject.toml
[tool.poetry]
name = "pychain"
version = "0.1.0"
description = "A simple blockchain API"
authors = ["Will Parks <parkswdev@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "0.12.3"
requests = "2.18.4"
black = "^20.8b1"
coloredlogs = "^15.0"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
pytest = "^4.6"
flake8 = "^3.8.4"
black = {version = "^20.8b1", allow-prereleases = true}


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"